payment-gateway/Data-Sources/requirements.txt
mysql-connector-python==8.0.33
faker==18.11.2
tqdm==4.65.0